+++ GWorkspace/Dialogs/StartAppWin.m
@@ -70,7 +70,7 @@ static NSString *nibName = @"StartAppWin";
 - (void)showWindowWithTitle:(NSString *)title
                     appName:(NSString *)appname
                   operation:(NSString *)operation
-               maxProgValue:(float)maxvalue
+               maxProgValue:(double)maxvalue
 {
   if (win) {
     [win setTitle: title];
@@ -87,7 +87,7 @@ static NSString *nibName = @"StartAppWin";
   }
 }
                  
-- (void)updateProgressBy:(float)incr
+- (void)updateProgressBy:(double)incr
 {
   [progInd incrementBy: incr];
 }
